TaxMenu
Added a new tax menu. Press T to bring up the tax menu, which allows the player to individually increase and decrease the individual rates (light res prop, light res inc, etc.).
These rates are not currently implemented in any other calculations, and the current "base tax rate" system needs to be removed in favor of the individual rates.